requirements.txt
feedparser==6.0.11
requests==2.32.3
APScheduler==3.10.4
python-dotenv==1.0.1
